Pull out the model into a separate file
